requirements.txt
asgiref==3.9.1
Django==5.2.5
django-parler==2.3
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
sqlparse==0.5.3
whitenoise==6.11.0
